requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
openpyxl
python-multipart
aiofiles
python-docx
